+++ Assets/Xbox Live/GameSave/GameSaveHelper.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: bdba0886bc35ae0478df61c0bf460e6d
+timeCreated: 1493753862
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
