+++ Assets/Xbox Live/GameSave/GameSaveStatus.cs
@@ -1,0 +1,22 @@
+﻿namespace Microsoft.Xbox.Services.ConnectedStorage
+{
+    public enum GameSaveStatus
+    {
+        Abort,
+        InvalidContainerName,
+        NoAccess,
+        OutOfLocalStorage,
+        UserCanceled,
+        UpdateTooBig,
+        QuotaExceeded,
+        ProvidedBufferTooSmall,
+        BlobNotFound,
+        NoXboxLiveInfo,
+        ContainerNotInSync,
+        ContainerSyncFailed,
+        UserHasNoXboxLiveInfo,
+        ObjectExpired,
+        Ok
+    }
+}
+
