+++ Assets/Xbox Live/GameSave/GameSaveUI.prefab.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: a39dcb3a34209ba4fa3abeb4260c0b38
+timeCreated: 1494361212
+licenseType: Free
+NativeFormatImporter:
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
