+++ Assets/Xbox Live/GameSave/README.txt
@@ -1,0 +1,3 @@
+To use Connected Storage in your game, you need to first install "Xbox Live Platform Extensions SDK" from: http://aka.ms/xblextsdk
+
+After installing the SDK, double click on the GameSave.unitypackage found at 'Assets\Xbox Live\Scripts\GameSave\' to import the scripts. To quickly test your integration, drag 'GameSaveUI' and drop it on any object you have in your scene.
