+++ Assets/Xbox Live/GameSave/README.txt.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2bf2dcae188656e4a9b8046fbc05ff14
+timeCreated: 1494349255
+licenseType: Free
+TextScriptImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
