--- Assets/Xbox Live/Prefabs/GameSaveUI.prefab.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 189e3689c66a9df40a6dedcd008cc9b6
-timeCreated: 1494350337
-licenseType: Free
-NativeFormatImporter:
-  mainObjectFileID: 100100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
