--- Assets/Xbox Live/Scripts/GameSave.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 33f8ec79399db26418d1dcd62797d23d
-folderAsset: yes
-timeCreated: 1493753842
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
