--- Assets/Xbox Live/Scripts/GameSave/GameSaveStatus.cs
@@ -1,22 +1,0 @@
-﻿namespace Microsoft.Xbox.Services.ConnectedStorage
-{
-    public enum GameSaveStatus
-    {
-        Abort,
-        InvalidContainerName,
-        NoAccess,
-        OutOfLocalStorage,
-        UserCanceled,
-        UpdateTooBig,
-        QuotaExceeded,
-        ProvidedBufferTooSmall,
-        BlobNotFound,
-        NoXboxLiveInfo,
-        ContainerNotInSync,
-        ContainerSyncFailed,
-        UserHasNoXboxLiveInfo,
-        ObjectExpired,
-        Ok
-    }
-}
-
