--- Assets/Xbox Live/Scripts/GameSave/GameSaveStatus.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 5b2e1a3d5eab4b640be10dcf203baf8c
-timeCreated: 1493756177
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
