--- Assets/Xbox Live/Scripts/GameSave/GameSaveUI.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 8734de4e9ece6cd4a9b12de5adab602e
-timeCreated: 1493753895
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
