--- Assets/Xbox Live/Scripts/GameSave/README.txt.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2de2110ac4eebd146b9c358f8b3c544e
-timeCreated: 1494023748
-licenseType: Free
-TextScriptImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
