--- Assets/Xbox Live/Scripts/GameSave/StorageContainerInfo.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 33e7125b08d89e24d82cf6a652026b55
-timeCreated: 1493756189
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
