--- Assets/Xbox Live/Scripts/GameSave/StorageContainerQueryResult.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: b6bbb4eac86bfce42868a4b83da98ce1
-timeCreated: 1493756203
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
